requirements.txt
asgiref==3.8.1
Django==4.2.1
sqlparse==0.5.0
tzdata==2024.1
psycopg2-binary==2.9.9
pillow==10.3.0
black==24.4.2
ipython==8.25.0
pytils==0.4.1
django-phonenumber-field[phonenumberslite]
python-dotenv